backend/requirements.txt
flask==2.3.3
Flask-CORS==4.0.0
flask-socketio==5.3.6
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==2.13.0
keras==2.13.0
scikit-learn==1.3.0
pandas==2.0.3
Pillow==10.0.0.1
requests==2.31.0
python-dotenv==1.0.0

# DeepFace library for emotion detection (modern alternative to FER)
deepface==0.0.79

# Optional: Advanced facial landmark detection (requires compilation)
# dlib==19.24.1
